allow blocTests to add more than one asynchronous event
The issue was that emitsExactly was listening to events too late. Since
Bloc uses a BehaviorSubject, if there is a delay between events the
listener would miss states emitted before the listener attaches. The
listener should start listening before any events are added so that all
states get recorded. I suspect this was the underlying problem with
[#724].