Update architecture.md
I changing the bulleted list of layers so that Presentation is on top and Data is on bottom. This fits the order that graphic above it (where UI is first). It also matches the text that follows, which says, "We're going to start at the lowest level layer [i.e., Data layer] and work our way up to the presentation layer."