Add package name for `events.throttle(duration)`
The `events.throttle(duration)` method depends on 'package:stream_transform/stream_transform.dart'. Personally I was following along the docs, and I had an error on the throttle method, I wasted around 30 minutes figuring out why the error occurred. Only when I went to the next I realized that there is a package I need to import. I know this change might seem unnecessary but trust me I think it adds some value to the documentation and saves the time of the reader.