feat(flutter_bloc): support listener current state
Allows setting a `startWithCurrentState` flag on `BlocListener` which
changes the behaviour to immediately call `listener` with the current
`state` of the bloc before calling again with subsequent emissions.

This allows the view layer to react to whatever the current state is
rather than only when it changes.

For example:

```dart
return BlocListener<AuthBloc, AuthState>(
  // Ensure the listener is checked immediately on creation with
  // current state in addition to future transitions.
  // If the user is somehow routed to this page when not logged in
  // they should be immediately kicked out, rather than only when
  // they transition from logged in to logged out.
  startWithCurrentState: true,
  listener: (context, state) {
    if(!state.isLoggedIn) {
      context.router.replace(const UnauthenticatedRoute());
    }
  },
  child: MyAuthenticatedPageContent(),
);
```

Closes #4347